reorg and cleanup
