+++ inc/constants.h
@@ -32,10 +32,9 @@ extern t_pagemap* g_pagemap;
 #define PREV_CHUNK(chunk, prev_size) ((t_chunk*)((char*)(chunk) - prev_size))
 #define TINY_CHUNK_OVERHEAD sizeof(size_t)
 
-// Alignment to ensure proper boundaries
-#define ALIGN_SIZE 8
-#define ALIGN_MASK (ALIGN_SIZE - 1)
-#define ALIGN(n) (((n) + ALIGN_MASK) & ~ALIGN_MASK)
+// Helper macros for cache_table
+#define FNV_OFFSET 14695981039346656037UL
+#define FNV_PRIME 1099511628211UL
 
 // =================== Shifters ===================
 
@@ -54,4 +53,3 @@ extern t_pagemap* g_pagemap;
 #define NUM_BINS 149
 
 #endif // CONSTANTS_H
-;
