+++ inc/main.h
@@ -10,6 +10,7 @@
 #include <math.h>
 #include <assert.h>
 #include <stdint.h>
+#include <string.h>
 
 #include "utils.h"
 #include "constants.h"
