+++ src/frontend_cache/fast_cache.c
@@ -3,13 +3,10 @@
 t_tiny_chunk** create_fast_cache(t_cache* cache) {
   t_tiny_chunk** fast_cache = (t_tiny_chunk**)CACHE_SHIFT(cache);
 
-  // printf("fast_cache pointer: %p\n", fast_cache);
   for (size_t i = 0; i < cache->fcache_size; ++i) {
     fast_cache[i] = NULL;
   }
 
-  // void* last_byte = (void*)MEMORY_SHIFT(cache, (cache->fcache_size * sizeof(t_tiny_chunk*)));
-  // printf("fast_cache end: %p\n", last_byte);
   return fast_cache;
 }
